Minecraft 1.21.5

+++ src/main/java/net/minecraftforge/common/extensions/IForgeEntity.java
@@ -98,7 +98,7 @@ public interface IForgeEntity extends ICapabilitySerializable<CompoundTag> {
      * @param fallDistance The fall distance
      * @return {@code true} if this entity can trample, {@code false} otherwise
      */
-    boolean canTrample(ServerLevel level, BlockState state, BlockPos pos, float fallDistance);
+    boolean canTrample(ServerLevel level, BlockState state, BlockPos pos, double fallDistance);
 
     /**
      * Returns The classification of this entity
